Fixed bug in clean -> dirty URL translations
The `url(archive/Y/m/d/)` constructor for daily archives is broken
because the monthly archive URL triggers a match before the daily
archive URL can be tested.

I’ve swapped the order of the URLs so that a match for daily archive is
tested first. I’ve also added a URL for the yearly archive.